fix:#263 ItemRequestControllerのupdateStatusメソッドのみを、Api/ItemRequestControllerへ移動、api.phpのuse分を修正
